feat: init bot

+++ src/config.rs
@@ -2,9 +2,9 @@ use std::fs;
 
 use serde::{Serialize, Deserialize};
 
-#[derive(Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Config {
-    pub app_id: u64,
+    pub app_id: i32,
     pub app_hash: String,
     pub telegram_bot_token: String,
     pub telegram_channel: String
@@ -18,7 +18,6 @@ impl Config {
             .expect("Isi file config dapat di parse oleh serde_json");
 
         drop(contents);
-
         deserialized
     }
 }
